Fix after review
ONE-DCO-1.0-Signed-off-by: Denis Krylov <krylovdenism@gmail.com>
