+++ compiler/circle2circle-dredd-recipe-test/test.lst
@@ -10,7 +10,7 @@
 ## TFLITE RECIPE
 
 Add(Net_InstanceNorm_001 PASS fuse_instnorm)
-Add(Net_InstanceNorm_003 PASS fuse_instnorm_v2)
+Add(Net_InstanceNorm_002 PASS fuse_instnorm)
 Add(BatchMatMulV2_000 PASS resolve_customop_batchmatmul)
 
 ## CIRCLE RECIPE
