+++ compiler/luci/pass/include/luci/CircleOptimizer.h
@@ -34,7 +34,6 @@ public:
     {
       FuseBCQ,
       FuseInstanceNorm,
-      FuseInstanceNormV2,
       ResolveCustomOpAdd,
       ResolveCustomOpBatchMatMul,
       QuantizeDequantizeWeights,
