+++ res/TensorFlowLiteRecipes/Net_InstanceNorm_002/test.rule
@@ -1,6 +1,8 @@
-# To check if custom op BatchMatMulV2 is converted to circle builtin op
+# To check if custom op InstanceNorm is converted to circle builtin op
 
 RULE    "VERIFY_FILE_FORMAT"      $(verify_file_format) '=' 1
 
 RULE    "INSTANCE_NORM_EXIST"     $(op_count INSTANCE_NORM) '=' 1
-RULE    "NO_RESHAPE"              $(op_count RESHAPE) '=' 0
+RULE    "RESHAPE_EXIST"           $(op_count RESHAPE) '=' 3
+RULE    "NO_ADD"                  $(op_count ADD) '=' 0
+RULE    "NO_MUL"                  $(op_count MUL) '=' 0
