--- res/TensorFlowLiteRecipes/Net_InstanceNorm_003/test.recipe
@@ -1,334 +1,0 @@
-operand {
-  name: "input_layer"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 32
-    dim: 32
-    dim: 8
-  }
-}
-operand {
-  name: "sequential/instance_normalization/stack"
-  type: INT32
-  shape {
-    dim: 5
-  }
-  filler {
-    tag: "explicit"
-    arg: "1"
-    arg: "32"
-    arg: "32"
-    arg: "8"
-    arg: "1"
-  }
-}
-operand {
-  name: "sequential/instance_normalization/Shape"
-  type: INT32
-  shape {
-    dim: 4
-  }
-  filler {
-    tag: "explicit"
-    arg: "1"
-    arg: "32"
-    arg: "32"
-    arg: "8"
-  }
-}
-operand {
-  name: "sequential/instance_normalization/Reshape_1"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 1
-    dim: 1
-    dim: 8
-    dim: 1
-  }
-  filler {
-    tag: "explicit"
-    arg: "0.0279993"
-    arg: "0.0496207"
-    arg: "-0.00456733"
-    arg: "0.0144199"
-    arg: "0.0331552"
-    arg: "0.0423247"
-    arg: "-0.00624703"
-    arg: "-0.00727953"
-  }
-}
-operand {
-  name: "sequential/instance_normalization/Reshape_2"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 1
-    dim: 1
-    dim: 8
-    dim: 1
-  }
-  filler {
-    tag: "explicit"
-    arg: "0.00417534"
-    arg: "-0.0214481"
-    arg: "-0.0176298"
-    arg: "-0.0352226"
-    arg: "0.0204345"
-    arg: "0.0374578"
-    arg: "0.0356524"
-    arg: "-0.0321159"
-  }
-}
-operand {
-  name: "sequential/instance_normalization/batchnorm/add/y"
-  type: FLOAT32
-  shape {
-  }
-  filler {
-    tag: "explicit"
-    arg: "0.001"
-  }
-}
-operand {
-  name: "sequential/instance_normalization/moments/variance/reduction_indices"
-  type: INT32
-  shape {
-    dim: 3
-  }
-  filler {
-    tag: "explicit"
-    arg: "1"
-    arg: "2"
-    arg: "4"
-  }
-}
-operand {
-  name: "sequential/instance_normalization/Reshape"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 32
-    dim: 32
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/moments/mean"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 1
-    dim: 1
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/moments/SquaredDifference"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 32
-    dim: 32
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/moments/variance"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 1
-    dim: 1
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/batchnorm/add"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 1
-    dim: 1
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/batchnorm/Rsqrt"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 1
-    dim: 1
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/batchnorm/mul"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 1
-    dim: 1
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/batchnorm/mul_1"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 32
-    dim: 32
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/batchnorm/mul_2"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 1
-    dim: 1
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/batchnorm/sub"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 1
-    dim: 1
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "sequential/instance_normalization/batchnorm/add_1"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 32
-    dim: 32
-    dim: 8
-    dim: 1
-  }
-}
-operand {
-  name: "Identity"
-  type: FLOAT32
-  shape {
-    dim: 1
-    dim: 32
-    dim: 32
-    dim: 8
-  }
-}
-operation {
-  type: "Reshape"
-  input: "input_layer"
-  input: "sequential/instance_normalization/stack"
-  output: "sequential/instance_normalization/Reshape"
-}
-operation {
-  type: "Mean"
-  input: "sequential/instance_normalization/Reshape"
-  input: "sequential/instance_normalization/moments/variance/reduction_indices"
-  output: "sequential/instance_normalization/moments/mean"
-  mean_options {
-    keep_dims: true
-  }
-}
-operation {
-  type: "SquaredDifference"
-  input: "sequential/instance_normalization/Reshape"
-  input: "sequential/instance_normalization/moments/mean"
-  output: "sequential/instance_normalization/moments/SquaredDifference"
-}
-operation {
-  type: "Mean"
-  input: "sequential/instance_normalization/moments/SquaredDifference"
-  input: "sequential/instance_normalization/moments/variance/reduction_indices"
-  output: "sequential/instance_normalization/moments/variance"
-  mean_options {
-    keep_dims: true
-  }
-}
-operation {
-  type: "Add"
-  input: "sequential/instance_normalization/moments/variance"
-  input: "sequential/instance_normalization/batchnorm/add/y"
-  output: "sequential/instance_normalization/batchnorm/add"
-  add_options {
-    activation: NONE
-  }
-}
-operation {
-  type: "Rsqrt"
-  input: "sequential/instance_normalization/batchnorm/add"
-  output: "sequential/instance_normalization/batchnorm/Rsqrt"
-}
-operation {
-  type: "Mul"
-  input: "sequential/instance_normalization/batchnorm/Rsqrt"
-  input: "sequential/instance_normalization/Reshape_1"
-  output: "sequential/instance_normalization/batchnorm/mul"
-  mul_options {
-    activation: NONE
-  }
-}
-operation {
-  type: "Mul"
-  input: "sequential/instance_normalization/Reshape"
-  input: "sequential/instance_normalization/batchnorm/mul"
-  output: "sequential/instance_normalization/batchnorm/mul_1"
-  mul_options {
-    activation: NONE
-  }
-}
-operation {
-  type: "Mul"
-  input: "sequential/instance_normalization/moments/mean"
-  input: "sequential/instance_normalization/batchnorm/mul"
-  output: "sequential/instance_normalization/batchnorm/mul_2"
-  mul_options {
-    activation: NONE
-  }
-}
-operation {
-  type: "Sub"
-  input: "sequential/instance_normalization/Reshape_2"
-  input: "sequential/instance_normalization/batchnorm/mul_2"
-  output: "sequential/instance_normalization/batchnorm/sub"
-  sub_options {
-    activation: NONE
-  }
-}
-operation {
-  type: "Add"
-  input: "sequential/instance_normalization/batchnorm/mul_1"
-  input: "sequential/instance_normalization/batchnorm/sub"
-  output: "sequential/instance_normalization/batchnorm/add_1"
-  add_options {
-    activation: NONE
-  }
-}
-operation {
-  type: "Reshape"
-  input: "sequential/instance_normalization/batchnorm/add_1"
-  input: "sequential/instance_normalization/Shape"
-  output: "Identity"
-}
-input: "input_layer"
-output: "Identity"
